Add files via upload
The objective of this project is to predict the bank loan default case with the help of various factors of customers. By achieving this goal, it would be possible to help accommodate in managing the customer for their loans on a daily basis, and providing better services to its customer.